Oracle.html
Selection Bug (Intentions & Spreads): My attempt to fix this with a simple click handler wasn't sufficient. Mobile browsers can be tricky. I'll implement a more forceful approach: when you click the label, the JavaScript will not only check the hidden radio button but also explicitly trigger its 'change' event. This should guarantee that the rest of the app logic (like enabling the 'Continue' buttons) recognizes the selection.

Stuck Spinner: This is likely happening because the part of the code that hides the spinner isn't being reached, possibly due to an error during the (simulated) data processing or display. I will move the "hide spinner" command into a finally block. This ensures the spinner is hidden regardless of whether the reading succeeds or fails, preventing it from getting stuck.

Version Number: Good call. I'll add a version comment to the top.

Signed-off-by: JustWingIt1011 <ho.richard@gmail.com>